Fix some bug for audit rules set.

+++ bin/hardening/8.5_ensure_permissions_on_all_logfiles.sh
@@ -41,7 +41,9 @@ apply () {
 	else
 		warn  "Permissions of all log files are not correctly configured! Set it"
 		chmod -R $PERMISS_SET $LOGDIR/*
-		rm $ERRPERFILELIST
+		if [ -r $ERRPERFILELIST ]; then
+			rm $ERRPERFILELIST
+		fi
 	fi
 }
